Eliminate macros by making Andex use a phantom type, Andexer
We now use a "phantom data" type parameter in Andex, to make the
generic type unique every time it's instantiated with a different
type. The type itself can actually be an empty enum, as it's not
instantiated.

AndexableArray is the generic wrapper for arrays, which we should be
using with a type alias. It gets the Andexer as a parameter, and
supports being indexed by it out-of-the-box, no macros required. And
the "phantom" type prevents cross-indexing, while still allowing more
than one AndexableArray for each Andex.